separating Play.page, MatchHistory, and Match into different folders. Sorting scss classes in separate scss files. Improving the organization of code.
