+++ src/App.scss
@@ -99,7 +99,7 @@ body {
     --penalty-word-yellow: #f5e856;
 
     background: var(--background-color);
-    --tertiary-background-color: #273649;
+    --tertiary-background-color: #253850;
     --green-adjusted: rgb(11, 59, 11);
   }
