+++ src/pages/Match.page/styles.scss
@@ -1,0 +1,44 @@
+.match {
+  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  padding: 10px;
+  .head {
+    margin-bottom: 8px;
+  }
+  .spectators {
+    background-color: var(--secondary-background-color);
+    padding: 5px;
+    border-radius: 5px;
+    margin-bottom: 8px;
+  }
+  .timer {
+    margin-bottom: 10px;
+    background-color: var(--secondary-background-color);
+    padding: 5px;
+    border-radius: 5px;
+  }
+  .list {
+    background-color: var(--secondary-background-color);
+    display: flex;
+    justify-content: space-between;
+    margin: 3px;
+    border-radius: 10px;
+    padding: 8px;
+    min-width: 35%;
+  }
+  .button {
+    background: var(--tertiary-background-color);
+    border-radius: 8px;
+    font-size: 20px;
+    margin: 10px;
+    &:hover {
+      filter: opacity(75%);
+    }
+  }
+  .active-background {
+    background-color: var(--tertiary-background-color) !important;
+  }
+}
