+++ src/pages/MatchList.page/styles.scss
@@ -1,0 +1,147 @@
+.history-main > div,
+.history-body > div {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+.history-main {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  .history-header {
+    padding: 10px;
+    border-radius: 10px;
+    background-color: var(--secondary-background-color);
+    margin-bottom: 15px;
+    .history-header-top {
+      display: flex;
+      justify-content: space-between;
+      width: 100%;
+      font-size: large;
+      margin-bottom: 10px;
+    }
+    .history-header-date {
+      opacity: 0.5;
+    }
+  }
+  .history-body {
+    .history-body-list {
+      width: 100%;
+      max-width: 650px;
+      .title {
+        font-size: large;
+        margin-bottom: 7px;
+      }
+      .card {
+        background-color: var(--secondary-background-color);
+        display: flex;
+        justify-content: space-between;
+        margin: 3px;
+        border-radius: 10px;
+        padding: 10px;
+        width: 125%;
+      }
+    }
+    .history-text {
+      margin-top: 10px;
+    }
+  }
+  .history-button {
+    .button {
+      font-size: 20px;
+      font-weight: 300;
+      border-radius: 8px;
+      margin-bottom: 10px;
+      padding: 2px 8px;
+      margin-top: 20px;
+      background-color: var(--tertiary-background-color);
+      &:hover {
+        filter: opacity(75%);
+      }
+    }
+  }
+}
+
+.match-body {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  //SearchBar
+  > input {
+    max-width: 960px;
+    align-self: center;
+  }
+
+  .match-body-main {
+    margin-top: 55px;
+    display: flex;
+    width: 100%;
+    max-width: 960px;
+    justify-content: space-between;
+    padding: 0 15px;
+    position: relative;
+
+    .match-body-main-list {
+      display: flex;
+      justify-content: start;
+      align-items: center;
+      flex-direction: column;
+      gap: 0.75rem;
+      flex-grow: 1;
+      .match-card {
+        padding: 0.5rem;
+        background-color: var(--secondary-background-color);
+        border-radius: 10px;
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        box-shadow:
+          0 4px 6px -1px rgb(0 0 0 / 0.1),
+          0 2px 4px -2px rgb(0 0 0 / 0.1);
+        min-height: 100px;
+        justify-content: start;
+        width: 80%;
+        min-width: 360px;
+        justify-content: space-between;
+
+        animation: slideInFromLeft 500ms ease-in-out forwards;
+
+        .head {
+          display: flex;
+          justify-content: space-between;
+        }
+        .match-card-bottom {
+          display: flex;
+          flex-direction: row;
+          justify-content: space-between;
+          .button {
+            font-weight: 300;
+            border-radius: 5px;
+            background-color: var(--tertiary-background-color);
+            &:hover {
+              filter: opacity(75%);
+            }
+          }
+          p {
+            opacity: 0.5;
+          }
+        }
+      }
+    }
+
+    .back {
+      position: absolute;
+      right: 0;
+      top: -45px;
+      .button {
+        font-size: 20px;
+        font-weight: 300;
+        border-radius: 8px;
+        padding: 3px 10px;
+      }
+    }
+  }
+}
